auth-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pyjwt==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
